contracts: Remove types and storage from the public interface
