+++ bin/node/cli/src/chain_spec.rs
@@ -295,10 +295,9 @@ pub fn testnet_genesis(
 			phantom: Default::default(),
 		},
 		pallet_contracts: ContractsConfig {
-			current_schedule: pallet_contracts::Schedule {
-				enable_println, // this should only be enabled on development chains
-				..Default::default()
-			},
+			// println should only be enabled on development chains
+			current_schedule: pallet_contracts::Schedule::default()
+				.enable_println(enable_println),
 		},
 		pallet_sudo: SudoConfig {
 			key: root_key,
